requirements.txt
aiogram==3.13.1
sqlalchemy==2.0.35
aiosqlite==0.20.0
alembic==1.13.3
python-dotenv==1.0.1
loguru==0.7.2
pillow==10.4.0
pytz==2024.2